Update kern.c
I rewrote the <data> if-statements in both kern_read_handle and kern_write_handle functions. Since pfmalloc returns NULL on error, I redid all if-statements that check the return value of pfmalloc to be if (<variable> == NULL) rather than if (!<variable>).
I renamed the variable <rp> in kern_read_handle function to <kReadPkt>.
I rewrote the if-statement for <kReadPkt> variable in kern_read_handle, to check if <ReadPkt> is invalid, rather than to check if it's valid and have the code inside of that if-statement.
I renamed the variable <wp> in kern_write_handle function to <kWritePkt>.
I rewrote the if-statement for <kWritePkt> variable in kern_write_handle, to check if <WritePkt> is invalid, rather than to check if it's valid and have the code inside of that if-statement.
I Commented the code of both kern_write_handle and kern_read_handle functions.